535 - refactored the multi scroller code to a simpler, easier to maintain version of its former self
