+++ src/pointerAction.js
@@ -9,7 +9,7 @@ import {
     TRIGGERS
 } from "./constants";
 import {observe, unobserve} from "./helpers/observer";
-import {armGlobalScroller, disarmGlobalScroller, didAnyContainerScrollDuringLastInterval} from "./helpers/globalScroller";
+import {createMultiScroller} from "./helpers/multiScroller";
 import {
     createDraggedElementFrom,
     decorateShadowEl,
@@ -89,11 +89,6 @@ function unregisterDropZone(dropZoneEl, type) {
 function watchDraggedElement() {
     printDebug(() => "watching dragged element");
     const dropZones = typeToDropZones.get(draggedElType);
-    // it is important that we don't have an interval that is faster than the flip duration because it can cause elements to jump bach and forth
-    const setIntervalMs = Math.max(...Array.from(dropZones.keys()).map(dz => dzToConfig.get(dz).dropAnimationDurationMs));
-    const observationIntervalMs = setIntervalMs === 0 ? DISABLED_OBSERVATION_INTERVAL_MS : Math.max(setIntervalMs, MIN_OBSERVATION_INTERVAL_MS); //if setintervalms is 0 it goes to 20, otherwise it is max between it and min observation.
-    const intervalMS = observationIntervalMs * 1.07;
-    armGlobalScroller(typeToDropZones.get(draggedElType), intervalMS);
 
     for (const dz of dropZones) {
         dz.addEventListener(DRAGGED_ENTERED_EVENT_NAME, handleDraggedEntered);
@@ -101,11 +96,15 @@ function watchDraggedElement() {
         dz.addEventListener(DRAGGED_OVER_INDEX_EVENT_NAME, handleDraggedIsOverIndex);
     }
     window.addEventListener(DRAGGED_LEFT_DOCUMENT_EVENT_NAME, handleDrop);
-    observe(draggedEl, dropZones, didAnyContainerScrollDuringLastInterval, intervalMS);
+
+    // it is important that we don't have an interval that is faster than the flip duration because it can cause elements to jump bach and forth
+    const setIntervalMs = Math.max(...Array.from(dropZones.keys()).map(dz => dzToConfig.get(dz).dropAnimationDurationMs));
+    const observationIntervalMs = setIntervalMs === 0 ? DISABLED_OBSERVATION_INTERVAL_MS : Math.max(setIntervalMs, MIN_OBSERVATION_INTERVAL_MS); // if setIntervalMs is 0 it goes to 20, otherwise it is max between it and min observation.
+    const multiScroller = createMultiScroller(dropZones, () => currentMousePosition);
+    observe(draggedEl, dropZones, observationIntervalMs * 1.07, multiScroller);
 }
 function unWatchDraggedElement() {
     printDebug(() => "unwatching dragged element");
-    disarmGlobalScroller();
     const dropZones = typeToDropZones.get(draggedElType);
     for (const dz of dropZones) {
         dz.removeEventListener(DRAGGED_ENTERED_EVENT_NAME, handleDraggedEntered);
